% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrt_pcr.R
\name{eds.test}
\alias{eds.test}
\title{eds.test}
\usage{
eds.test(data, Reference_gene, Measured_genes, Experimental_conditions = NULL,
  Replicate_id = "_")
}
\arguments{
\item{data}{This is a list from the eds.summary function}

\item{Reference_gene}{The name of your reference gene that you are measuring relative expression against
as a character}

\item{Measured_genes}{A vector of the genes you are measuing as characters}

\item{Experimental_conditions}{If you have multiple experimental conditions that you want to compare this
is a character vector of those conditions. By default it is NULL}
}
\value{
This is the function that produces the results. Essentially it uses the ddCT
method to produce a list of dataframes, one for each gene. Each dataframe contains the
mean, sd and se for each sample. There is also a "Gene and Condition" tag to make plotting easy.
}
\description{
eds.test
}
\examples{
Measured_genes <- c("GOI_1", "GOI_2")
eds.test(test2, "PDF2", Measured_genes, Experimental_conditions = Experimental_conditions)
}

