% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_time.R
\name{flr.hist.graph}
\alias{flr.hist.graph}
\title{flr.hist.graph}
\usage{
flr.hist.graph(data, meas = "fl", new.titles = NULL)
}
\arguments{
\item{data}{This is a a flr.summary object used to draw the graphs.}

\item{meas}{The measurement of flowering time. This is by default "fl" for days to flowering but could also be "nd" for nodes at the time of flowering.}

\item{new.titles}{By default this function labels each facet with the entry in the Genotype column of the flr.summary object. This can be changed by inputting a named vector of names which map the genotypes to the new names.}
}
\value{
This is a function for drawing a histogram of flowering time measurments from a screen. It takes a flr.summary object as input and outputs a graph for each line highlighting homozygotes. At this stage the Mutant column needs to be either "WT" or "Homo"
}
\description{
flr.hist.graph
}
\examples{
new_labs <- c(`NF5076` = "NF5076 - LD+V", `R108` = "R108 - LD+V")
}

