% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_time.R
\name{flr.hist}
\alias{flr.hist}
\title{flr.hist}
\usage{
flr.hist(data, new.titles = NULL)
}
\arguments{
\item{data}{This is a a flr.summary object used to draw the graphs}

\item{new.titles}{By default this function labels each facet with the entry in the Genotype column of the flr.summary object. This can be changed by inputting a named vector of names which map the genotypes to the new names.}
}
\value{
This is a wrapper function for flr.hist.graph in that it calls it twice. Once for days to flower and once for nodes at the time of flowering. It then arranges them.
}
\description{
flr.hist
}
\examples{
new_labs <- c(`NF5076` = "NF5076 - LD+V", `R108` = "R108 - LD+V")
flr.hist(res, new.titles = new_labs)
}
