% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_time.R
\name{flr.plot}
\alias{flr.plot}
\title{flr.plot}
\usage{
flr.plot(data)
}
\arguments{
\item{data}{This is a dataframe from flr.summary}
}
\value{
This function plots flowering time data. If you a just comparing one line and its wild-type control it
 will colour and plot these side by side. If there are multiple conditions then it will colour these and facet by genotype.
}
\description{
flr.plot
}
\examples{
flr.plot(data)
}
