% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{g_colours}
\alias{g_colours}
\title{Geoffrey's Colours}
\format{A vector of length 10 containing hexcodes.}
\source{
These colours are taken from the wesanderson library 
    and consist of the Darjeeling and Rushmore sets.
}
\usage{
g_colours
}
\description{
These are the colours I use in my plots.
}
\keyword{datasets}
