% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{g_colours2}
\alias{g_colours2}
\title{Geoffrey's Colours 2}
\format{A vector of length 5 containing hexcodes.}
\source{
These colours are taken from the Metropolis Beamer Theme 
    which can be found here: github.com/matze/mtheme.
}
\usage{
g_colours2
}
\description{
These are the colours I use in my plots.
}
\keyword{datasets}

