% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{ggplotRegression}
\alias{ggplotRegression}
\title{ggplotRegression}
\usage{
ggplotRegression(fit)
}
\arguments{
\item{fit}{is a linear model from the lm function}
}
\value{
This plots a linear model with ggplot2. It is a function written by Susan 
Johnston (https://susanejohnston.wordpress.com/2012/08/09/a-quick-and-easy-function-to-plot-lm-results-in-r/).
}
\description{
ggplotRegression
}

