% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{ggtimeline}
\alias{ggtimeline}
\title{ggtimeline}
\usage{
ggtimeline(x, begin, end)
}
\arguments{
\item{x}{is a data.frame with three columns}

\item{begin}{is the date the timeline begins in the form "dd/mm/yyyy"}

\item{end}{is the date the timeline ends in the form "dd/mm/yyyy"}
}
\value{
This function draws a timeline given a data.frame with three columns. 
These columns should correspond to activity, start and end. The start and end columns 
should be dates in the form of "dd/mm/yyyy". You also need to specific when 
you want the time table to begin and end.
}
\description{
ggtimeline
}
\examples{
timeline(timeline_dates, "1/1/2015", "1/3/2018")
}
