% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{moe}
\alias{moe}
\title{moe}
\usage{
moe(data, alpha = 0.05)
}
\arguments{
\item{data}{A vector of numeric type}

\item{alpha}{The significance level (default = 0.05)

This calculates a margin of error (half a confidence interval) with the default being 95%.}
}
\description{
moe
}
