% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{my_theme}
\alias{my_theme}
\title{my_theme}
\usage{
my_theme(base_size = 12, base_family = "sans")
}
\value{
This is the theme I use for my plots. Simple and based on theme_wsj().
Use it with theme_set or as a parameter in a ggplot call.
}
\description{
my_theme
}
\examples{
theme_set(my_theme)
}
