% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pal}
\alias{pal}
\title{pal}
\usage{
pal(col, border = "light gray", ...)
}
\arguments{
\item{col}{A vector of hexcodes}
}
\value{
This function displays a colour pallete, you can change the colour of the border but its probably not neccessary.
Written by Achim Zeileis (in the colorspace package apparently) but I found it here:
http://www.r-bloggers.com/the-paul-tol-21-color-salute/
}
\description{
pal
}
\examples{
pal(g_colours)
}

