% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pickRandomRows}
\alias{pickRandomRows}
\title{pickRandomRows}
\usage{
pickRandomRows(df, numberOfRows = 10)
}
\arguments{
\item{df}{A dataframe}

\item{numberOfRows}{Number of rows to sample

This function randomly selects rows from a data frame. I got it from here/; http://www.markhneedham.com/blog/2014/11/26/r-dplyr-select-random-rows-from-a-data-frame/}
}
\description{
pickRandomRows
}
