% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrt_pcr.R
\name{read.eds}
\alias{read.eds}
\title{read.eds}
\usage{
read.eds(path, in_file = "Results")
}
\arguments{
\item{path}{This is a character string to point you to the file}

\item{in_file}{This is the type of input file. By default it is "Results" for Ct results
but you can also set it to "Melt" for melting curve data.}
}
\value{
This function imports csv files from a eds file produced by the Quant 5
RT-qPCR machine into R as a dataframe. It does this by cutting off the header, selecting the
relevant columns and ordering the rows.
}
\description{
read.eds
}
\examples{
read.eds(Path)
}

