% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read.nano}
\alias{read.nano}
\title{read.nano}
\usage{
read.nano(Path, n_samples)
}
\arguments{
\item{Path}{The path to the Nanophotometer output file}

\item{n_samples}{The number of Nanophotometer measurements (i.e. if you did one twice count it twice)

This reads the excel file produced by the Nanophotometer and outputs a markdown table.}
}
\description{
read.nano
}
\examples{
Path = "~/Documents/Nanophotometer/21_7_16_.xlsx"
read_nano(Path, 7)
}

