% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrt_pcr.R
\name{read.sds}
\alias{read.sds}
\title{read.sds}
\usage{
read.sds(path, col = c("Well", "Sample", "Detector", "Reporter", "Ct"))
}
\arguments{
\item{path}{This is a character string to point you to the file}

\item{col}{These are the columns you want to keep, by default it selects
"Well", "Sample",  "Detector", "Reporter" and "Ct". If you want to change
it it takes a character vector}
}
\value{
This function imports SDS 2.4 files from a RT-qPCR machine into
R as a dataframe. It does this by cutting off the header, selecting the
relevant columns and ordering the rows.
}
\description{
read.sds
}
\examples{
read.sds(Path)
}
