% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_time.R
\name{flr.plot}
\alias{flr.plot}
\title{flr.plot}
\usage{
flr.plot(data)
}
\arguments{
\item{data}{This is a dataframe from flr.summary}
}
\value{
This function plots flowering time data
}
\description{
flr.plot
}
\examples{
flr.plot(data)
}

